[*] Some fixes.
git-svn-id: https://flush.svn.sourceforge.net/svnroot/flush/trunk@29 21f5d154-ea94-4097-a85b-94d12d4401c9

+++ src/daemon_settings.hpp
@@ -89,7 +89,7 @@
 					Auto_clean_type		max_ratio_type;
 
 					/// Максимальное количество раздающих торрентов.
-					int					max_seeding_torrents;
+					ssize_t				max_seeding_torrents;
 					Auto_clean_type		max_seeding_torrents_type;
